[bug-fix] 补上makefile文件，同时修复以下警告。 * RRNode.h中出现的变量声明顺序与构造函数变量初始化顺序不一致。 * FpgaTile.h中assert中出现的int与size_t类型不匹配比较问题。

+++ lab3/FpgaTile.h
@@ -58,12 +58,12 @@ public:
     }
 
     RRNode &getVWire(int idx) {
-        assert(idx < vWires.size());
+        assert((size_t) idx < vWires.size());
         return *(vWires[idx]);
     }
 
     RRNode &getHWire(int idx) {
-        assert(idx < hWires.size());
+        assert((size_t) idx < hWires.size());
         return *(hWires[idx]);
     }
